[Move/#16] diaryfeed -> diary 모듈명 변경
