+++ .gitignore
@@ -1,4 +1,5 @@
 *.iml
+*.idea
 .gradle
 /local.properties
 /.idea/caches
