+++ feature/diary/proguard-rules.pro
