+++ feature/diary/src/main/AndroidManifest.xml
