+++ feature/diary/src/main/java/com/memozi/diary/DiaryScreen.kt
@@ -1,4 +1,4 @@
-package com.memozi.diaryfeed
+package com.memozi.diary
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
