--- feature/diaryfeed/.gitignore
@@ -1,3 +1,0 @@
-# Created by https://www.toptal.com/developers/gitignore/api/android,androidstudio,kotlin
-# Edit at https://www.toptal.com/developers/gitignore?templates=android,androidstudio,kotlin
-/build
